[DROOLS-3401] Moved Fluent API to kie-internal
