+++ kie-internal/src/main/java/org/kie/internal/builder/fluent/DataSourceFluent.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2016 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,9 +14,19 @@
  * limitations under the License.
  */
 
-package org.kie.api.runtime.builder;
+package org.kie.internal.builder.fluent;
 
-public interface TimeFluent <T> {
-    T after(long duration);
-    T relativeAfter(long duration);
+import org.kie.api.runtime.rule.DataSource;
+
+/**
+ * See {@link DataSource}
+ */
+public interface DataSourceFluent<E, U extends RuleUnitFluent> {
+
+    DataSourceFluent<E, U> addBinding(String dataSourceName);
+
+    DataSourceFluent<E, U> insert(E object);
+
+    U buildDataSource();
+
 }
