+++ kie-internal/src/main/java/org/kie/internal/builder/fluent/KieSessionFluent.java
@@ -14,29 +14,12 @@
  * limitations under the License.
  */
 
-package org.kie.api.runtime.builder;
+package org.kie.internal.builder.fluent;
 
-import org.kie.api.runtime.rule.FactHandle;
-import org.kie.api.runtime.rule.RuleRuntime;
-import org.kie.api.runtime.rule.StatefulRuleSession;
-
-/**
- * See {@link RuleRuntime} and {@link StatefulRuleSession}
- */
-public interface RuleFluent<T, U> {
-
-    T fireAllRules();
-
-    T setGlobal( String identifier, Object object );
-
-    T getGlobal(String identifier);
-
-    T insert(Object object);
-
-    T update( FactHandle handle, Object object );
-
-    T delete(FactHandle handle);
-
-    U dispose();
+public interface KieSessionFluent
+    extends RuleFluent<KieSessionFluent, ExecutableBuilder>,
+    ProcessFluent<KieSessionFluent, ExecutableBuilder>,
+    ContextFluent<KieSessionFluent, ExecutableBuilder>,
+    TimeFluent<KieSessionFluent> {
 
 }
