+++ kie-internal/src/main/java/org/kie/internal/builder/fluent/RuleUnitFluent.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.api.runtime.builder;
+package org.kie.internal.builder.fluent;
 
 import java.util.function.Function;
 import java.util.function.Supplier;
