+++ kie-internal/src/main/java/org/kie/internal/builder/fluent/TaskFluent.java
@@ -13,9 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.kie.internal.builder.fluent;
 
-package org.kie.api.runtime.builder;
+import org.kie.api.task.TaskService;
 
-public enum Scope {
-    REQUEST, CONVERSATION, APPLICATION;
+/**
+ * See {@link TaskService}
+ *
+ */
+public interface TaskFluent<T> {
+
 }
